config: order ports list for clarity
Iso is down to 600MB due to new base/kernel configs.  One of
the biggest offenders now are some build tools that are not
viable for stripping in FreeBSD 10.0, as well as python27.
If the image must be smaller we need to rethink a bit, if not
I guess this is fine for a long time now.