scripts: step towards system upgrades
We need to keep track of new and old binaries and directories.
This script will show removed and added files, but really only
the removed files are interesting as they must be stripped from
the system to not cause any harm (think vulnerable libraries
like OpenSSL from base we want to get rid of soon).