build/nano: first pass for embedded images; fixes #5
Builds an image that loads in bhyve, but might still need some more
tweaking.  We don't have two slices.  I am not sure if they'll be
back soon.  Not sure if these images are ready for release as well,
but at least now we have the script for providing initial snapshots.