build: stagify remaining pieces
This will not taint the build system except for the final
sets and image dirs.  Helps in keeping the build system
clean and that it's easier to inspect the build step.

While there, expand chroot to take a subdir argument as
some build steps need the STAGEDIR root for more than a
single purpose.