build/nano: complete revamp leaving NanoBSD behind
Images are clamped to just under 4G with a single slice
in a hybrid manner of vm/serial images.  This should
provide wide compatibility and fit any 4G card, and can
then be expanded to whatever we really have using growfs.

Smaller images are possible too, but not recommended due
to the fact that firmware upgrades need to be downloaded
and extracted.  We could make 2G the new default, but
future size requirements and a steady 4G image seem like
a more favourable goal in the long term.