build/skim: add edit mode to upstream sync
A process has been established over the last few weeks to manually
alter the commits before committing them.  Now it's time to weave
this into the script to save a little time in the future:

Run git-checkout(1) in patch mode in order to discard chunks from
the commit, and skip the commit if the result is actually empty.