build/dvd: the script formerly known as cdrom
After much discussion and questions, we move to a DVD-ish nomenclature
to indicate that images are likely oversized for a CD. In any case,
the Virtual Machines don't care about this, but users deciding to try
OPNsense might.

Now is the time as the next images are likely going to be for 17.7.