build/distfiles: fix fetch-recursive + FLAVOR fail
FreeBSD says "don't do that", but it's clearly an error on the
framework part.  To make the code more robust, kill the usage
of the fetch-recursive target and just do what we do in skim by
using all-depends-list.

When we have that list, we iterate over it with a small caveat:

We strip the flavor again to avoid breakage in the future which
could be pretty likely given that the functionality is very new
and not yet found to be used in its full potential.