build: remove spurious WRKDIR change
It was needed at some point in the past but hasn't been useful
for a good while now.  Recently it broke because the plist test
is now run prior to 'package' target and while it may be fixed
there by creating the directory this code using a custom WRKDIR
is the bigger bug so good-bye. :)