config: to discuss...
Lua version used is 5.3.

Rust is 139 Mb download to update sets.  Sets are 500Mb currently
so that adds just shy of 30% download space required and an additional
unpack space of the same amount for major upgrades for absolutely
no benefit.

I guess the requirements come from OPNids so we will find a way,
but that currently is not it because the footprint is too high.  :(