build: sign our mtree set files for later verification
Right now we trust the mtree file since we can't guarantee
consistency after unpack.  This way we will be able to warn
on unsigned mtree file or when the signature is wrong meaning
somebody meddled with the mtree file.

Should work for both base and kernel.

PR: https://github.com/opnsense/core/issues/3473