build: do not enter 20.7.r description territory on 20.1
Due to an issue with pkg upgrade -f[1] and lack of a better approach
we will avoid version parity beween 20.1 and 20.7 development branches
by omitting 20.1 from using the RC tag for versioning.

This allows us to force an upgrade always even though the actual
version is going backwards.  It doesn't really matter then because
we seek full transition into the new packages that have been published
and not try to outsmart the upstream repo's efforts to keep everything
in sync.

[1] https://github.com/freebsd/pkg/issues/1806