config: add cups package
This request has been voiced a couple of times over the years.
With some mileage behind us it would make sense to enhance the
system with a print server that we don't offer any plugins for
or at least that is not planned from the core team side.

As with everything: use at your own risk.  ;)