build: change CONFIG to CONFIGDIR/CONFIGDIRPREFIX approach
Always require build.conf since /dev/null does not work anymore
due to omission of embedded language versions.

The top view CONFIGDIR is not the low level CONFIGDIR in the
build scripts, but for lack of refactoring noise and a catchy
naming scheme let's just leave it where it is.