build: make sure the build can cope with SUFFIX=-business
We do not want to install globs as the matching is very fuzzy and
also we no longer do.  To make sure prevent use of globs in
install_packages so we can catch this in nightly builds if it
ever happens.

While here, allow to set a VERSION for core builds which ignores
the tag-version-derivation.  It is useful for prerelease building.

Last but not least ADDITIONS should no longer look for SUFFIX so
if you want to build an image with SUFFIX=-devel and use a development
plugin overwrite ADDITIONS accordingly, e.g. ADDTIONS=os-dyndns-devel