build/ports: it seems that pkg -N sometimes wants to bootstrap
The /usr/sbin/pkg supports -N in the right way, but /usr/local/sbin/pkg -N
which is being invoked if /usr/local/sbin/pkg exists has other semantics
that sometimes makes this work and sometimes it doesn't.  The manual page
mentions a weird workaround, but in the end if we assume that base pkg
execs local pkg we can almost certainly assume that it is bootstrapped and
usable.