build/serial: rebuild UEFI on FreeBSD 13
The .efifat files have been deleted from the distribution so they
need to be handrolled.  That's alright, but then mkimg doesn't work
on FreeBSD 13 because our STAGEDIR contains a ":" character.
Sidestep the issue by using a relative path inside STAGEDIR, but
also hoping that file names with ":" are not becoming second-class.

This should be compatible with 21.7 builds although we likely do not
need new images.

While here also convert dvd image with others to be done later.