config: scrub non-generic ident from generic SMP kernels
ident SMP-XXX is where XXX designates a device type to carry a
base/kernel as "sticky" which was introduced for 32bit ARM where
every device required a different kernel (or boot loader address).

For the generic 64bit ARM kernel and soon to be merged RSS variant
using this makes no sense and complicates testing for users.