make: remove standard addition of os-dyndns
According to e0e08e9a5b0 this was asked about in the forum back
then, but the plugin hasn't gotten much improvement over the years
and to use it you need an Internet connection which you can also
use to install the plugin.

On a restore you can already sync all missing plugins now which is
also helping in this case.