build/vm: support bundling a ZFS image; closes #272
Flip the order of root and swap to allow VM image to grow.
Hook this up in extras.conf and also retrofit a serial console
for being able to test and debug in bhyve (build/boot).

Omit the use of a GPT root label in the ZFS case to deal with
devices names in the way that bsdinstall sets it up.  It is
not all that important for ZFS anyway.

The setup likely needs further testing and tweaking to make
sure we emulate bsdinstall zfsboot script properly to avoid
future problems.