config: enable curl CA_BUNDLE which is no longer the defaut
We might be able to get away with it as well but need to ensure that
${LOCALBASE}/etc/ssl/cert.pem is the authoritative file for CA info.

FreeBSD nowadays has a separate stash somewhere but we still feed it
from ca_root_nss package.  Merging the two is an option, but not having
ca_root_nss anymore is not a good idea.  Luckily, core already depends
on ca_root_nss separately so that this will be a quick POC once 23.1
is out and can be released around 23.1.x.