composite/hotfix: switch default hotfixing to partial core/plugins build
I have been unintentionally scrubbing packages sets in the past which
is not a lot of fun so make the default hotfix mode a little less coarse.

While here the biggest issue of the hotfixing is not being able to see
package messages (not errors, these are shown as the build aborts) so
collect them in a variable to show at the end if all was building ok.
This is only relevant for multi-step hotfixing between core/plugins.