build/plugins: defer plugin builds which have plugin dependencies
We did avoid this for as long as possible and through plugins.conf.local
that was automatic for os-OPNProxy before but now we need to put a bit
more effort in in order to not fail the bulk plugins build because of
os-squid not being ready (yet).  It's also a complication of os-squid
no longer being in core but there are still situations like this in plugins
and core that require this fix.