build: install the pkgs version file for the image
opnsense-update handles this file on major upgrades, but on a fresh
install the version is shown as N/A because the file is not yet created.

The best time to create it is when we install packages.  The approximation
is using the PRODUCT_CORE package version as the concepts between packages
set and core package version may differ.  Also for a distribution we use
the core package version to create the image name as well.