config: replace pymongo shadow copy with correct port
Recent python glue changes broke our shadow copy which used to be
in the FreeBSD ports tree some time ago before it was renamed.  As
such it never received updates so eventually it had to break.

Adding this dates back to 2020 and there doesn't appear to be a
visible downstream user so maybe it can be removed but for now just
make the build work again by using the proper port.